# swagger-full.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "图书馆座位预约系统 API",
        "version": "1.0.0",
...
    },
    "servers": [
        {
            "url": "http://localhost:3000",
            "description": "开发环境"
        },
        {
            "url": "http://localhost:3001",
            "description": "测试环境"
        },
        {
            "url": "http://localhost:3002",
            "description": "UAT 环境"
        },
        {
            "url": "http://localhost:3003",
            "description": "生产环境"
        }
    ],
...
    "paths": {
...
        "/api/feedback": {
            "post": {
                "tags": ["💬 反馈系统 (Feedback)"],
                "summary": "提交反馈",
                "security": [
                    {
                        "bearerAuth": []
                    }
                ],
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "type": "object",
                                "required": ["typeId", "title", "description"],
                                "properties": {
                                    "typeId": {
                                        "type": "integer",
                                        "enum": [1, 2, 3, 4],
                                        "description": "【必填】反馈类型 (1:功能建议，2:问题上报，3:投诉建议，4:其他)"
                                    },
                                    "urgencyId": {
                                        "type": "integer",
                                        "enum": [1, 2, 3, 4],
                                        "description": "【可选】紧急程度 (1:低，2:中，3:高，4:紧急)"
                                    },
                                    "title": {
                                        "type": "string",
                                        "description": "【必填】反馈标题"
                                    },
                                    "description": {
                                        "type": "string",
                                        "description": "【必填】反馈描述"
                                    },
                                    "contact": {
                                        "type": "string",
                                        "description": "【可选】联系方式"
                                    },
                                    "images": {
                                        "type": "array",
                                        "items": {
                                            "type": "string",
                                            "format": "uri"
                                        },
                                        "description": "【可选】图片 URL 列表"
                                    }
                                }
                            }
                        }
                    }
                },
                "responses": {
                    "200": {
                        "description": "提交成功",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "type": "object",
                                    "properties": {
                                        "success": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "data": {
                                            "type": "object",
                                            "properties": {
                                                "id": {
                                                    "type": "string"
                                                },
                                                "typeId": {
                                                    "type": "integer"
                                                },
                                                "urgencyId": {
                                                    "type": "integer"
                                                },
                                                "title": {
                                                    "type": "string"
                                                },
                                                "description": {
                                                    "type": "string"
                                                },
                                                "status": {
                                                    "type": "integer",
                                                    "enum": [1, 2, 3, 4],
                                                    "description": "反馈状态"
                                                },
                                                "createdAt": {
                                                    "type": "string",
                                                    "format": "date-time"
                                                }
                                            }
                                        }
                                    }
                                }
                            }
                        }
                    },
                    "400": {
                        "description": "参数错误"
                    }
                }
            }
        },
...
    },
    "components": {
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "scheme": "bearer",
                "bearerFormat": "JWT",
                "description": "JWT Token 认证，格式：Bearer <token>"
            }
        },
...
    }
}